requirements.txt
streamlit
transformers
pdfplumber
torch
Pillow
